refactored old user modal into a new page
