+++ client/src/pages/lab_management/users/AccessCheckCard.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { Button, Card, CardActions  } from "@mui/material";
-import { AccessCheckExtraInfo, GET_USER } from "./UserModal";
+import { AccessCheckExtraInfo, GET_USER } from "./UserPage";
 import { gql, useMutation } from "@apollo/client";
 import AuditLogEntity from "../audit_logs/AuditLogEntity";
