+++ client/src/pages/lab_management/users/ManageUserArchive.tsx
@@ -1,6 +1,6 @@
 import { gql, useMutation } from "@apollo/client";
 import { Stack, ToggleButton, ToggleButtonGroup, Typography } from "@mui/material";
-import { GET_USER } from "./UserModal";
+import { GET_USER } from "./UserPage";
 import { isManager } from "../../../common/PrivilegeUtils";
 import { useCurrentUser } from "../../../common/CurrentUserProvider";
