+++ client/src/pages/lab_management/users/RestrictionCard.tsx
@@ -1,6 +1,6 @@
 import gql from "graphql-tag";
 import { useCurrentUser } from "../../../common/CurrentUserProvider";
-import { GET_USER, Restriction } from "./UserModal";
+import { GET_USER, Restriction } from "./UserPage";
 import { Button, Card, CardActions, CardContent, CardHeader, Stack, Typography, useTheme } from "@mui/material";
 import { useMutation } from "@apollo/client";
 import { format, parseISO } from "date-fns";
