+++ client/src/pages/lab_management/users/UserPage.tsx
@@ -202,13 +202,8 @@ const DELETE_PASSED_MODULE = gql`
   }
 `;
 
-interface UserModalProps {
-  selectedUserID: string;
-  onClose: () => void;
-}
-
-export default function UserModal({ selectedUserID, onClose }: UserModalProps) {
-  const { makerspaceID } = useParams<{ makerspaceID: string }>();
+export default function UserModal() {
+  const { makerspaceID, userID} = useParams<{ makerspaceID: string, userID: string }>();
   const navigate = useNavigate();
   const currentUser = useCurrentUser();
   const isMobile = useIsMobile();
@@ -223,15 +218,15 @@ export default function UserModal({ selectedUserID, onClose }: UserModalProps) {
   const [createHold] = useMutation(CREATE_HOLD);
   const [createRestriction] = useMutation(CREATE_RESTRICTION);
   const [setNotesMutation, setNotesResult] = useMutation(SET_NOTES);
-  const [refreshCheck, refreshCheckResult] = useMutation(REFRESH_CHECKS, { variables: { userID: selectedUserID }, refetchQueries: [{ query: GET_USER, variables: { id: selectedUserID } }] });
-  const [createCheck] = useMutation(CREATE_CHECK, { refetchQueries: [{ query: GET_USER, variables: { id: selectedUserID } }] });
-  const [deleteTrainingHold] = useMutation(DELETE_TRAINING_HOLD, { refetchQueries: [{ query: GET_USER, variables: { id: selectedUserID } }] });
-  const [deletePassedModule] = useMutation(DELETE_PASSED_MODULE, { refetchQueries: [{ query: GET_USER, variables: { id: selectedUserID } }] });
+  const [refreshCheck, refreshCheckResult] = useMutation(REFRESH_CHECKS, { variables: { userID: userID }, refetchQueries: [{ query: GET_USER, variables: { id: userID } }] });
+  const [createCheck] = useMutation(CREATE_CHECK, { refetchQueries: [{ query: GET_USER, variables: { id: userID } }] });
+  const [deleteTrainingHold] = useMutation(DELETE_TRAINING_HOLD, { refetchQueries: [{ query: GET_USER, variables: { id: userID } }] });
+  const [deletePassedModule] = useMutation(DELETE_PASSED_MODULE, { refetchQueries: [{ query: GET_USER, variables: { id: userID } }] });
   const getMakerspacesResult = useQuery(GET_FULL_MAKERSPACES);
 
   useEffect(() => {
-    if (selectedUserID) getUser({ variables: { id: selectedUserID } });
-  }, [selectedUserID, getUser]);
+    if (userID) getUser({ variables: { id: userID } });
+  }, [userID, getUser]);
 
   const handlePlaceHoldClicked = () => {
     const description = window.prompt("Enter hold description:");
@@ -245,7 +240,7 @@ export default function UserModal({ selectedUserID, onClose }: UserModalProps) {
 
     createHold({
       variables: { userID: getUserResult.data.user.id, description },
-      refetchQueries: [{ query: GET_USER, variables: { id: selectedUserID } }],
+      refetchQueries: [{ query: GET_USER, variables: { id: userID } }],
     });
   };
 
@@ -265,7 +260,7 @@ export default function UserModal({ selectedUserID, onClose }: UserModalProps) {
 
     createRestriction({
       variables: { userID: getUserResult.data.user.id, makerspaceID: restrictionMakerspace, reason: reason },
-      refetchQueries: [{ query: GET_USER, variables: { id: selectedUserID } }]
+      refetchQueries: [{ query: GET_USER, variables: { id: userID } }]
     });
   }
 
@@ -275,7 +270,7 @@ export default function UserModal({ selectedUserID, onClose }: UserModalProps) {
 
   function handleCheckCreate() {
     if (!newCheckEquipmentID) return;
-    createCheck({ variables: { userID: selectedUserID, equipmentID: newCheckEquipmentID } });
+    createCheck({ variables: { userID: userID, equipmentID: newCheckEquipmentID } });
     setOpenCreateCheckDialouge(false);
   }
 
@@ -284,7 +279,7 @@ export default function UserModal({ selectedUserID, onClose }: UserModalProps) {
   };
 
   return (
-    <PrettyModal open={!!selectedUserID} onClose={onClose} width={isMobile ? 300 : 1600}>
+    <Box margin="10px 25px">
       <RequestWrapper2
         result={getUserResult}
         render={({ user }) => {
@@ -315,7 +310,7 @@ export default function UserModal({ selectedUserID, onClose }: UserModalProps) {
                     <Typography>{user.pronouns}</Typography>
                   </Stack>
                 </Stack>
-                <IconButton color="error" onClick={onClose} sx={{ width: "34px", height: "34px" }}>
+                <IconButton color="error" onClick={() =>navigate(`/makerspace/${makerspaceID}/people`)} sx={{ width: "34px", height: "34px" }}>
                   <CloseIcon />
                 </IconButton>
               </Stack>
@@ -397,7 +392,7 @@ export default function UserModal({ selectedUserID, onClose }: UserModalProps) {
                               <IconButton
                                 color="error"
                                 disabled={!isStaffFor(currentUser, module.moduleID ?? -1)}
-                                onClick={() => deletePassedModule({ variables: { userID: selectedUserID, moduleID: module.moduleID } })}
+                                onClick={() => deletePassedModule({ variables: { userID: userID, moduleID: module.moduleID } })}
                               >
                                 <DeleteIcon />
                               </IconButton>
@@ -531,14 +526,14 @@ export default function UserModal({ selectedUserID, onClose }: UserModalProps) {
                     placeholder="Notes"
                     value={notes}
                     onChange={handleNotesChanged}
-                    onSubmit={() => setNotesMutation({ variables: { userID: selectedUserID, notes: notes } })}
+                    onSubmit={() => setNotesMutation({ variables: { userID: userID, notes: notes } })}
                     multiline
                     minRows={2}
                   />
                   <Button
                     variant="contained"
                     loading={setNotesResult.loading}
-                    onClick={() => setNotesMutation({ variables: { userID: selectedUserID, notes: notes } })}
+                    onClick={() => setNotesMutation({ variables: { userID: userID, notes: notes } })}
                     sx={{ alignSelf: "flex-end" }}
                   >
                     Update Notes
@@ -549,6 +544,6 @@ export default function UserModal({ selectedUserID, onClose }: UserModalProps) {
           );
         }}
       />
-    </PrettyModal >
+    </Box>
   );
 }
